requirements.txt
selenium==4.19.0
webdriver-manager==4.0.1
PyQRCode==1.2.1
openpyxl==3.1.2
pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1